Additional gcc 12 and C++20 fixes.

+++ unicode/biditest.C
@@ -21,8 +21,9 @@ int main(int argc, char **argv)
 		char32_t c=atoi(argv[1]);
 		unicode_bidi_bracket_type_t bt;
 
-		std::cout << unicode_bidi_mirror(c) << " "
-			  << unicode_bidi_bracket_type(c, &bt) << " ";
+		std::cout << (unsigned long)unicode_bidi_mirror(c) << " "
+			  << (unsigned long)unicode_bidi_bracket_type(c, &bt)
+			  << " ";
 		std::cout << (char)bt << std::endl;
 		exit (0);
 	}
